requirements.txt
openai==1.7.0
python-dotenv==1.0.0
websockets==11.0.3